Fix AudioFrameV3Wrapper float32 consistency and buffer syntax error
Changed AudioFrameV3Wrapper to consistently use float32 for the public FLTP payload: data_array type, set_data parameter and static_cast, get_data return type, and the py::class_ binding.
Fixed extra closing parenthesis in py::array(py::buffer_info(...)) call that broke compilation. sizeof(float) now matches the float format descriptor for correct stride calculation.

+++ src/main.cpp
@@ -256,10 +256,10 @@ struct AudioFrameV3Wrapper {
   void set_data(py::array_t<float> arr) {
     data_array = std::move(arr);
     auto info = data_array.request();
-    inner.p_data = static_cast<float *>(info.ptr);
+    inner.p_data = static_cast<uint8_t *>(info.ptr);
     inner.no_channels = info.shape[0];
     inner.no_samples = info.shape[1];
-    inner.channel_stride_in_bytes = info.strides[0];
+    inner.channel_stride_in_bytes = static_cast<int>(info.strides[0]);
   }
   py::array get_data() const {
     if (!inner.p_data)
